lab1: improvements
1. Add read_size function
2. Some renaming
3. Replace substr function with custom implementation

+++ lab1/src/utils.h
@@ -12,6 +12,6 @@ size_t find(const std::string& haystack, const char needle, size_t pos);
 
 size_t find_first_not_of(const std::string& haystack, const char needle, size_t pos);
 
-void insert_in_str(std::string& dest, const std::string& src, int index);
+void insert_in_str(std::string& dest, const std::string& src, size_t index);
 
 #endif /* #ifndef UTILS_H */
